+++ package.json
@@ -6,8 +6,8 @@
   "repository": "github:hashgraph/hedera-hardhat-hethers",
   "author": "hashgraph",
   "license": "MIT",
-  "main": "internal/index.js",
-  "types": "internal/index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "keywords": [
     "hedera",
     "smart-contracts",
@@ -25,41 +25,43 @@
     "clean": "node_modules/.bin/rimraf dist internal types *.{d.ts,js}{,.map} build-test tsconfig.tsbuildinfo"
   },
   "files": [
-    "dist/src/",
-    "src/",
-    "internal/",
-    "types/",
-    "*.d.ts",
-    "*.d.ts.map",
-    "*.js",
-    "*.js.map",
+    "dist/",
     "LICENSE",
     "README.md"
   ],
   "devDependencies": {
     "@hashgraph/hedera-local": "^1.1.0",
     "@hashgraph/hethers": "^1.1.0",
-    "@types/chai": "^4.2.0",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^12.0.0",
-    "@typescript-eslint/eslint-plugin": "4.29.2",
-    "@typescript-eslint/parser": "4.29.2",
-    "axios": "^0.26.1",
+
+    "@types/chai": "^4.1.7",
+    "@types/find-up": "^4.0.0",
+    "@types/fs-extra": "^5.0.4",
+    "@types/js-yaml": "^3.12.5",
+    "@types/mocha": "^5.2.5",
+    "@types/node": "^8.10.38",
     "chai": "^4.2.0",
-    "dotenv": "^16.0.0",
-    "eslint": "^7.29.0",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-import": "2.24.1",
-    "eslint-plugin-prettier": "3.4.0",
-    "hardhat": "^2.0.0",
-    "mocha": "^9.2.2",
-    "prettier": "2.4.1",
-    "rimraf": "^3.0.2",
+    "hardhat": "^2.10.1",
+    "mocha": "^10.0.0",
+    "prettier": "^1.17.0",
+    "source-map-support": "^0.5.12",
     "ts-node": "^8.1.0",
-    "typescript": "~4.5.2"
+    "tslint": "^6.1.3",
+    "tslint-config-prettier": "^1.18.0",
+    "tslint-plugin-prettier": "^2.0.1",
+    "typescript": "^4.7.4",
+    "axios": "^0.26.1",
+    "dotenv": "^16.0.0"
   },
   "peerDependencies": {
     "@hashgraph/hethers": "^1.1.0",
-    "hardhat": "^2.0.0"
+    "hardhat": "^2.10.1"
+  },
+  "dependencies": {
+    "@ethersproject/bignumber": "^5.6.2",
+    "@nomiclabs/hardhat-ethers": "^2.0.6",
+    "ethers": "^5.6.8",
+    "fs-extra": "^9.0.1",
+    "hardhat-deploy": "^0.11.10",
+    "js-yaml": "^3.14.0"
   }
 }
