--- src/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "./../tsconfig-common.json",
-  "compilerOptions": {
-    "outDir": "../",
-    "rootDirs": [
-      "./"
-    ],
-    "composite": true
-  },
-  "include": ["./**/*.ts"],
-  "exclude": []
-}
