+++ test/fixture-projects/hardhat-project/hardhat.config.js
@@ -18,7 +18,7 @@
  * ‍
  */
 
-require("../../../src/internal");
+require("../../../src");
 const path = require('path');
 require('dotenv').config({path: path.resolve(__dirname, '../../../.env')});
