+++ test/index.ts
@@ -24,9 +24,9 @@ import type { hethers } from "@hashgraph/hethers";
 import { hethers as hethersObj } from "@hashgraph/hethers";
 import { NomicLabsHardhatPluginError } from "hardhat/plugins";
 import { Artifact } from "hardhat/types";
-import { HethersProviderWrapper } from "../src/internal/hethers-provider-wrapper";
+import { HethersProviderWrapper } from "../src/hethers-provider-wrapper";
 import { useEnvironment } from "./helpers";
-import { SignerWithAddress } from "../src/internal/signers";
+import { SignerWithAddress } from "../src/signers";
 import path from 'path';
 import dotenv from "dotenv";
 dotenv.config({path: path.resolve(__dirname, '../.env')});
