--- tsconfig-common.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2017",
-    "module": "commonjs",
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "strict": true,
-    "esModuleInterop": true,
-    "typeRoots": ["../../node_modules/@types", "./@types"],
-    "noEmitOnError": true,
-    "skipDefaultLibCheck": true,
-    "skipLibCheck": true
-  }
-}
