+++ tslint.json
@@ -1,0 +1,33 @@
+{
+  "extends": [
+    "tslint:latest",
+    "tslint-plugin-prettier",
+    "tslint-config-prettier"
+  ],
+  "rules": {
+    "prettier": true,
+    "object-literal-sort-keys": false,
+    "no-submodule-imports": false,
+    "interface-name": false,
+    "max-classes-per-file": false,
+    "no-empty": false,
+    "no-console": false,
+    "only-arrow-functions": false,
+    "variable-name": [
+      true,
+      "check-format",
+      "allow-leading-underscore",
+      "allow-pascal-case"
+    ],
+    "ordered-imports": [
+      true,
+      {
+        "grouped-imports": true,
+        "import-sources-order": "case-insensitive"
+      }
+    ],
+    "no-floating-promises": true,
+    "prefer-conditional-expression": false,
+    "no-implicit-dependencies": true
+  }
+}
